+++ tests/getPluginConfig-test.js
@@ -6,7 +6,7 @@ import getPluginConfig from '../src/getPluginConfig'
 
 describe('getPluginConfig()', () => {
   it('scans package.json for nwb-* dependencies and imports them', () => {
-    let config = getPluginConfig(path.join(__dirname, 'fixtures/plugins'))
+    let config = getPluginConfig({cwd: path.join(__dirname, 'fixtures/plugins')})
     expect(config).toEqual({
       cssPreprocessors: {
         fake: {